wip: cucumber generates both gherkin and ruby files
Previously, generating cucumber needed to run hiptest-publisher twice:
 - the first time for feature files and step definitions ruby file
 - the second time for the actionwords (using --language=ruby
   --actionwords-only)

Now all files can be generated in one go.

How does it work: the language configuration file, in
lib/config/<language>[-<framework>].conf defines the files to generate
and the templates (ie the language) to use for them. The Renderer is
called for each node with a context describing how the node should be
rendered.