wip: make distinction between the two generated code categories
Some generated code should not be modified, and is perfect as-is. They
are generally generated from the tests/scenarios. It is named 'test code'.
The other generated code are stubs that needs to be implemented. They are
generally generated from the actionwords. It is named 'actionwords
stubs'.

The distinction is made because some languages need to export tests
and/or actionwords multiple times. Code generated from actionwords can
be totally reusable as-is: for example, in cucumber export, the step
definitions are generated from the actionwords. So it is false to assume
that any code generated from actionwords is to be implemented, that's
why a broader definition has been added: there is the test code on one part,
that should not be modified, and the actionwords stubs on the other part,
that should be implemented.