Autoguess the framework and fail if config not found for it
Before, you could issue --language=java --framework=junit4 and it will
happily continue, but it would have dismissed the framework. Now it
fails if the given framework could not be found.

Because framework can be given, some configuration files have been
renamed to include the framework name: ruby.conf was renamed as
ruby-rspec.conf

Also, as there is no ruby.conf file anymore, the framework is guessed
from the given language so we can find the config file for the
language/framework.