refactor rendering tests: Better mimic real behavior
We now supply the node to create the NodeRenderingContext instance, this
way the path is provided during tests, and this will be useful when
dealing with folders.

Having the `path` in context in tests also mean we do not need the `test_name`
anymore. It was here to provide the test name for Java and C# because it
could not be computed due to the absence of `path`