Output URL when it is used with --verbose
Before it was output before checking the command line arguments. Outputing
url could trigger the fetch of the test runs list before the command line
arguments are checked, that's why it has been done at the moment where
the http requests are done.

Also the output is not always immediate to prevent mixing output messages.