publisher can now fetch test run from old Hiptest installations
The api /publication/<token>/test_runs can be absent from Hiptest
on-premises installations, and hiptest-publisher uses this api to get
the list of available test runs since 0.13.0.

This route returns a 404. It was interpreted as if the token was
unknown. Now it continues and does not check if the test-run-id exists.