Correctly handle --no-uids option. Fixes #40
Options starting with "no-" are treated specially in optparse, it means
that the option is disabled, whereas we would want it to turn true. So I
changed the option to `uids` with a default value of `true` so that
`no-uids` turns it off.

I had to change our helper functions so that it reads the default
options set in our CliOptions class.