fix issue #43: added enhanced-overriden-template
This change should be backwards compatible.  It allows a someone overriding the templates to use the same directory structure that hip test does.  It allows for use of different actionwords.hbs in different context (step_defintions vs actionwords).

The only issue I found which may or may not be an easy of use design flaw is that in my use case I needed to copy the original cucumber/java/actionwords.hbs into my override folder because the wrong template was being picked up.  This is not caused by my change but by the search path design.