Spock: Actionword should return true
Spocks expect-block defines an automatic assertion-behavior, thus requiring a statement to return a boolean expression.
Since Actionwords can be reused in any sense it happens that an Actionword is used within an expect-block, hence the Actionword must return true.

Hiptest-Publisher should honor this behavior by generating true as the last statement in the stub. This alone will cause compilation errors when throwing UnsupportedOperationExceptions, thus requiring a change to the _steps template as well which should rather generate a failing assert (including message)