Output dataset uid for Scenario Outline, ref #75
With some gherkin based languages, the scenario outline title can
contain placeholders and they will be replaced during execution by the
junit formatter. See
this unit test for reference:
https://github.com/cucumber/cucumber-jvm/blob/0adf6694cb9006056a457511d8e9703d9b360682/core/src/test/java/cucumber/runtime/formatter/PrettyFormatterTest.java#L85-L116 :
The `<name>` inside the outline title gets replaced by the actual value
from the examples.

We use this functionality to get the dataset uid output in the junit xml for
each test executed.

This works for:

- cucumber-java
  - cucumber-js
  - behave

This does not work for others, but some output the whole example row so
we get the dataset uid anyway (ruby and specflow). The only problematic one
is Behat as it neither supports placeholders in scenario outline title or
outputs the example row.