Stop adding test uids by default, add them when test run is used
There is no point at exporting an extra column with uids when they can't
be fetched because no test run has been specified. So now the default is
to not have uids exported unless `--test-run-id` or `--test-run-name` is
used.

Of course this default behavior can be overriden by specifying `--uids`
or `--no-uids` on command line or in config file.

This will give cleaner feature files on CucumberStudio side.